+++ test/fixture/mock-generator/callable-type-hint/supported.php
@@ -2,4 +2,4 @@
 
 $message = 'Requires callable type hint support.';
 
-return $detector->isSupported('parameter.type.callable');
+return $detector->isSupported('type.callable');