+++ test/fixture/mock-generator/date-time-hhvm/builder.php
@@ -1,5 +1,0 @@
-<?php
-
-$builder = $factory->create('DateTime');
-
-return $builder->named('MockGeneratorDateTimeHhvm');