+++ test/fixture/mock-generator/iterable-type-hint/builder.php
@@ -1,0 +1,13 @@
+<?php
+
+$builder = $factory->create(
+    array(
+        'methodA' => function (
+            $phonySelf,
+            iterable $first,
+            iterable $second = null
+        ) {},
+    )
+);
+
+return $builder->named('MockGeneratorIterableTypeHint');