+++ test/fixture/mock-generator/iterable-type-hint/supported.php
@@ -1,0 +1,5 @@
+<?php
+
+$message = 'Requires iterable type hint support.';
+
+return $detector->isSupported('type.iterable');