+++ test/fixture/mock-generator/nullable-types/builder.php
@@ -1,0 +1,15 @@
+<?php
+
+$builder = $factory->create(
+    array(
+        'Eloquent\Phony\Test\TestInterfaceWithNullableTypes',
+        array(
+            'static customStaticMethod' =>
+                function (?string $string, ?stdClass $object) : ?TestClassA {},
+            'customMethod' =>
+                function (?string $string, ?stdClass $object) : ?TestClassA {},
+        ),
+    )
+);
+
+return $builder->named('MockGeneratorNullableTypes');