+++ test/fixture/mock-generator/nullable-types/supported.php
@@ -1,0 +1,5 @@
+<?php
+
+$message = 'Requires nullable types.';
+
+return $detector->isSupported('type.nullable');