+++ test/fixture/mock-generator/return-type-magic-call/builder.php
@@ -1,0 +1,11 @@
+<?php
+
+$builder = $factory->create(
+    array(
+        'static __callStatic' =>
+            function ($name, array $arguments) : \stdClass {},
+        '__call' => function ($name, array $arguments) : \stdClass {},
+    )
+);
+
+return $builder->named('MockGeneratorReturnTypeMagicCall');