+++ test/fixture/mock-generator/return-type-void/builder.php
@@ -1,0 +1,12 @@
+<?php
+
+$builder = $factory->create(
+    array(
+        'Eloquent\Phony\Test\TestInterfaceWithVoidReturnType',
+        array(
+            'customMethod' => function () : void {},
+        ),
+    )
+);
+
+return $builder->named('MockGeneratorVoidReturnType');