+++ test/fixture/mock-generator/return-type-void/supported.php
@@ -1,0 +1,5 @@
+<?php
+
+$message = 'Requires void return type declarations.';
+
+return $detector->isSupported('type.void');